Craig was right! Moved all PdfFilter implementations to the new files PdfFiltersPrivate.h and PdfFiltersPrivate.cpp. As all filters are created using PdfFilterFactory the implementations should not be accessible in a public header file.
Additionally added the virtual methods PdfFilter::CanEncode and PdfFilter::CanDecode which return wether a filter is able to 
encode or decode data. This was necessary as the RLE and LZW filter only decode data and are not able to encode data. Without this 
methods any user calling the encode methods of these filters will get an UnsupportedFilter exception without having any other way 
to check if encoding is possible with this filter.



git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@336 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5