Introduce an OPTIONAL class that parses a content stream (as read from PdfContentsTokenizer) and organizes it into a tree based on nesting of q/Q operators etc. It's intended for use in PoDoFoBrowser, and for content stream editing and manipulation.
The actual internal list of known operators is presently tiny, but it can round-trip
streams with unknown operators fine - they just won't affect the structure properly.
They'll still have any operands bound to them, but if they're a stateful/nested pair
that won't affect the tree structure correctly.

This class uses the Boost Graph Library to do its job. It is exposed to users of the
class, and it is intended that users directly manipulate the graph though the Boost
graph algorithms, visitors, etc. If Boost is not available, the class is not built
and is preprocessed out of the headers.


git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@740 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5