Various fixes while working with valgrind and callgrind.
FIXED: podofoimgextract displays the correct number of successfully extracted images now
FIXED: S M A L L optimization in PdfObject, the often called copy constructor creates
       the copy itself instead of calling the assignment operator, which results in several
       1000 less calls to PdfVariant() and PdfVariant::Clear()
FIXED: A memory leak in PdfStream
FIXED: Allocate TCIMapKey in PdfDictionary::Clear() only if the dictionary is not empty

I did lots of testing with valgrind today and was very happy to see that I was not able to detect a memory leak. The one found occurred only in the 
case of an exception and was easy to fix. Good job, guys! :)



git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@929 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5