Fix declaration of operator<< for PoDoFo::PdfString
Since PdfString is in the PoDoFo namespace, the operator<< for it must
be in the same namespace as well, otherwise it is not found. In
particular, operator<<(std::ostream&) is needed by cppunit as a way to
get the string representation of an arbitrary type, when using
CPPUNIT_ASSERT_EQUAL() on instances of it.

This used to work with GCC until 11 because of a buggy behaviour.
GCC 12 fixed it [1], causing this test to fail to build with it.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51577