PdfPageCollection: Stop trying to preserve document page structure tree, simplified caching
A combination of weird caching strategy and the attempt to preserve the
document page structure tree is causing fist iteration of all pages in the document
to be quadratic time. It's also causing performance issues when editing the
document structure in #85. ISO 32000-2:2020, 7.7.3.2 Page tree nodes says:
"PDF processors shall not be required to preserve the existing structure of the page tree".
This means that we can flatten the structure upon editing (copy pages inheritable attributes,
remove intermediate /Pages nodes), allowing all read write operations on
/Pages to happen at maximum linear time. Also:

- Remove a lot of hardly mainteneable code. PdfPageTreeCache is fully removed
and everything is done inside PdfPageCollection;
- Added PdfPageCollection::CreatePagesAt as requested in #86
    of hardly mainteneable code. Also PdfPageTreeCache is fully removed
    and everything is done inside PdfPageCollection