PdfIndirectObjectList: Removed a lookup when replacing/pushing an object into the structure
Uses the set<T>::insert(pos, ...) overload that has "amortized constant [complexity]
if the insertion happens in the position just before pos, logarithmic in the size of
the container otherwise". The given hint position iterator must be a valid one,
so an iterator that was used to extract a node can't be re-used.

See https://en.cppreference.com/w/cpp/container/set/insert