PdfIndirectObjectList: Fixed big performance regression (?) in lookup object with reference
std::lower_bound, which use dates back to 0.9.x, may have been logarithmic complexity
when used with std::vector, but it's documented to be linear time when used with containers
that don't have random access like std::set, which was pushed in 0.10. In fact, about
std::lower_bound cppreference[1] says: "the number of comparisons performed is logarithmic
in the distance between first and last (at most log 2(last - first) + O(1) comparisons). However,
for non-LegacyRandomAccessIterators, the number of iterator increments is linear".

The only fix is avoid using std::lower_bound and migrate the container to use heterogeneous
lookup for objects and references.

Fixes #108

[1] https://en.cppreference.com/w/cpp/algorithm/lower_bound#Complexity