charconv_compat: Fixed compilation in mingw < 12
For some reason it was found the STL library support in mingw 11.2 is slightly different
than other distributions of gcc 11 (typically linux distributions), and causes compilation
errors specific to <charconv> support. Everything seems aligned again MinGW 12, so
a clause looking specifically for __MINGW32__ (which is present also in x86_64 builds of
MinGW) is added testing for __GNUC__ < 12. Should fix #125