PdfFilter: Completely moved the class to private folder
This further reduces the public API surface. PdfFilter was
already an implementation detail and not exposed in any header,
and the user is not supposed to be able to inherit/create
filters directly. This turned to be necessary as some linkers
may not like classes with private constructors that are not used
in any place. It actually caused undefined error with GNU ld
linked binaries in some circumstances, for non obvious reason
as the symbols were never used externally