PdfParser/PdfXRefStreamParserObject: Better handle overflows in XRef sections/streams
It introduce for the first time google numerics in PoDoFo, and should provide a better approach to handle overflows in general and for #173.

Also clean errors in XRef sections/streams: negative values are InvalidXRef or InvalidXRefStream, but overflows or value bigger than internal limits are now ValueOutOfRange, because the section/streams are actually valid accordingly to modern PDF specifications.