PdfFreeObjectList: Clarified the use and the documentation of m_ObjectCount
It does not make sense to start at 1, like assuming the trailer is always an object
inside the PDF. The trailer with cross reference sections is not really indirect,
while XRef streams would be counted a second time with that logic. It also
didn't match at all with the previous documentation stating that GetObjectCount()
was the highest object number in the document. Now it's exactly like that:
GetObjectCount() corresponds to the highest object number ever used
in the document, and never decreases