PdfCommon: Addded SetMaxRecursionDepth(). Enlarged MaxRecursionDepthDefault to 450.
This effectively fixes the sample file in #183, which is read correctly
in Acrobat Reader. I briefly mesaured the stack in Windows and at
maximum it was ~150kb (Release build, ~420KB in Debug build)
with MaxRecursionDepthDefault = 450 and that file. We should better
measure the stack again, in different platforms, and/or make PdfParser
to not parse trailers with recursive functions