FileStreamDevice: now uses again C stdio for better performance
It was determined that fstream is a wrapper around <cstdio> in all relevant implementations
(GNU libstdc++, LLVM libc++, and also microsft/STL[1]). So we rollback FileStreamDevice
to use C stdio again to remove and indirection. Also many I/O functions in C++ STL
threats EOF as an error and sometimes it's quite hard to handle EOF witout manual
error clearing or other tricks[2][3][4], which in end look hacky and inefficient.

[1] https://github.com/microsoft/STL/blob/ad80eb79ba4953e7529d15b1bc8d0b540150bf7d/stl/inc/fstream#L780
[2] https://github.com/podofo/podofo/blob/57207ed2ddc12b38dc2d9d005539599e23a152be/src/podofo/private/PdfDeclarationsPrivate.cpp#L780
[3] https://github.com/podofo/podofo/blob/57207ed2ddc12b38dc2d9d005539599e23a152be/src/podofo/auxiliary/StreamDevice.cpp#L18
[4] https://github.com/podofo/podofo/blob/57207ed2ddc12b38dc2d9d005539599e23a152be/src/podofo/auxiliary/StreamDevice.cpp#L40