Improved creation of podofo-config.cmake and add podofo-config-version.cmake
It emerged from comments in[1][2] that current podofo-config.cmake version is not
really working to . Guide at[3] was followed. Also added public headers to the target
with install target_sources(... FILE_SET ...) as documented in [4], which were absent
from exports anyway in podofo-config.cmake as well. This increases the cmake
requirement to 3.23. Initial export patches were introduced by vcpkg maintainers,
as described here[5]

[1] https://github.com/xournalpp/xournalpp/pull/6364#issuecomment-2858805083
[2] https://github.com/xournalpp/xournalpp/pull/6364#issuecomment-2862755066
[3] https://cmake.org/cmake/help/latest/guide/tutorial/Adding%20Export%20Configuration.html
[4] https://cmake.org/cmake/help/latest/command/target_sources.html#file-sets
[5] https://github.com/microsoft/vcpkg/pull/31584#issuecomment-1612729778