PdfParser: Fixed stack overflow parsing documents with many XRef stream updates
We basically remove on more try-catch-rethrow that is not playing well with RecursionGuard.
In this case the catch block was checking for PdfErrorCode::InvalidNumber exception
and not throwing in this case, but with improvements in PdfParserObject reading this
should not make sense anymore. It has always been a fragile solution anyway.

It has been reported in the ML https://www.mail-archive.com/podofo-users@lists.sourceforge.net/msg04995.html