PdfIndirectObjectList: Optimize object insertions by computing a proper hint iterator
The iterator must be retrieved with "set::lower_bound" and verified
for same object number. In this way the iterator can be used as a
proper hint in all cases. With "set::find" it returns and end iterator
if it's not an exact match, nullyfying its usefulness as an hint in case
of failed match